requirements.txt
aiohttp==3.7.4.post0
arrow==1.1.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
flake8==3.9.2
greenlet==1.1.0
idna==2.10
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.18
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
